fix: address more nits
